+++ R/install-url.r
@@ -11,7 +11,7 @@
 #' @param config additional configuration argument (e.g. proxy,
 #'   authentication) passed on to \code{\link[httr]{GET}}.
 #' @param before_install a function that can modify the contents of the package
-#'   source directory prior to installation. Accepts two parameters: bundle (the 
+#'   source directory prior to installation. Accepts two parameters: bundle (the
 #'   full path to the downloaded package zip file) and pkg_path (the path where
 #'   the package source is staged prior to installation)
 #' @param ... Other arguments passed on to \code{\link{install}}.