+++ NEWS.md
@@ -1,5 +1,8 @@
 # devtools 1.4.1.99
 
+* No longer checks for `pdftex` when building vignettes, as many modern
+  vignettes don't use it (#398).
+
 * `release()` does a better job of opening your email client if you're inside
   of Rstudio (#433).