+++ NEWS.md
@@ -1,5 +1,10 @@
 # devtools 1.4.1.99
 
+* `install_github()` now defaults to `dependencies = TRUE` so you definitely
+  get all the packages you need to build from source.
+
+* `install_deps(TRUE)` now includes packages listed in `VignetteBuilder` (#396)
+
 * No longer checks for `pdflatex` when building vignettes, as many modern
   vignettes don't use it (#398).